fix: mark self attendance from profile page
